envs_ecoClimate Obtain ecoClimate variables

download ecoClimate variables. See www.ecoclimate.org.

Usage

envs_ecoClimate(bcAOGCM, bcScenario, ecoClimSel, logger = NULL)

Arguments

bcAOGCM

Name of the Atmospheric and Oceanic Global Circulation Model. Options are: "CCSM", "CNRM", "MIROC", "FGOALS", "GISS", "IPSL","MRI", "MPI"

bcScenario

Select the temporal scenario that you want to download. Options are: "LGM" (21,000 years ago), "Holo" (6,000 years ago), "Present", "Future 2.6" (rcp 2.6), "Future 4.5" (rcp 4.5), "Future 6" (rcp 6), "Future 8.5" (rcp 8.5)

ecoClimSel

Numeric vector with list of variables to select.

logger

Stores all notification messages to be displayed in the Log Window of Wallace GUI. Insert the logger reactive list here for running in shiny, otherwise leave the default NULL.

Value

A rasterStack of selected variables

Details

This function is called by the module envs to download ecoClimate variables from www.ecoclimate.org. The variables to be downloaded are selected by the user with bcSel and the resolution is fixed to 0.5 degrees. This function currently gets variables from Dropbox and the process takes significantly more time than for other datasets. It returns a rasterStack of selected variables.

Author

Sara Varela <sara_varela@yahoo.com>

Jamie M. Kass <jamie.m.kass@gmail.com>

Gonzalo E. Pinilla-Buitrago <gepinillab@gmail.com>

Examples

bcAOGCM <- "CCSM"
bcScenario <- "LGM"
ecoClimSel <- c(1,2,3)
if (FALSE) { # \dontrun{
varsEcoClimate <- envs_ecoClimate(bcAOGCM, bcScenario, ecoClimSel)
} # }