model_bioclim Generate Bioclim model

The function generates a BIOCLIM model using ENMeval 2.0

Usage

model_bioclim(occs, bg, user.grp, bgMsk, logger = NULL, spN = NULL)

Arguments

occs

data frame of cleaned or processed occurrences obtained from components occs: Obtain occurrence data or, poccs: Process occurrence data.

bg

coordinates of background points to be used for modeling.

user.grp

a list of two vectors containing group assignments for occurrences (occs.grp) and background points (bg.grp).

bgMsk

a RasterStack or a RasterBrick of environmental layers cropped and masked to match the provided background extent.

logger

Stores all notification messages to be displayed in the Log Window of Wallace GUI. Insert the logger reactive list here for running in shiny, otherwise leave the default NULL.

spN

character. Species name to be used for all logger messages.

Value

Function returns an ENMevaluate object with all the evaluated models and a selection of appropriate fields.

Details

The function generates a model in ENMeval using a user provided partition of occurrences from previous components in the GUI.

See also

Author

Jamie M. Kass <jkass@gradcenter.cuny.edu>

Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>

Examples

if (FALSE) {
envs <- envs_userEnvs(rasPath = list.files(system.file("extdata/wc",
                                           package = "wallace"),
                      pattern = ".tif$", full.names = TRUE),
                      rasName = list.files(system.file("extdata/wc",
                                           package = "wallace"),
                      pattern = ".tif$", full.names = FALSE))
occs <- read.csv(system.file("extdata/Bassaricyon_alleni.csv",
                 package = "wallace"))
bg <- read.csv(system.file("extdata/Bassaricyon_alleni_bgPoints.csv",
               package = "wallace"))
partblock <- part_partitionOccs(occs, bg, method = 'block')
m <- model_bioclim(occs, bg, partblock, envs)
}