occs_queryDb Query online database for species occurrence records.

Queries a given database for occurrence data on the provided species

Usage

occs_queryDb(
  spNames,
  occDb,
  occNum = NULL,
  doCitations = FALSE,
  gbifUser = NULL,
  gbifEmail = NULL,
  gbifPW = NULL,
  RmUncertain = FALSE,
  logger = NULL
)

Arguments

spNames

character. Species Latin name, with format "Genus species".

occDb

character. Biodiversity database to query; current choices are "gbif", "vertnet", and "BIEN"

occNum

numeric. Maximum number of occurrence records to return

doCitations

logical. Set TRUE to use `occCite` to get a complete list of original data sources in a citable format

gbifUser

specify only if using `occCite` with GBIF to get a complete list of original data sources in a citable format. This, as well as `gbifEmail` and `gbifPW` are constraints imposed by GBIF to obtain the complete set of metadata associated with occurrence records and is not stored or used by `wallace` for any other purposes.

gbifEmail

specify only if using `occCite` with GBIF to get a complete list of original data sources in a citable format.

gbifPW

specify only if using `occCite` with GBIF to get a complete list of original data sources in a citable format.

RmUncertain

specify if occurrences without uncertainty information should be removed (default is FALSE)

logger

Stores all notification messages to be displayed in the Log Window of Wallace GUI. Insert the logger reactive list here for running in shiny, otherwise leave the default NULL

Value

list of lists one list per species with occurrence records. Each individual species list with appropriate fields for analysis

Details

This function is called by the module occs_queryDb to query a database for species occurrence records, subset to only those records with coordinates, remove records with duplicate coordinates, and select some columns with fields appropriate to studies in biogeography.

Author

Jamie Kass <jamie.m.kass@gmail.com>

Gonzalo E. Pinilla-Buitrago <gepinillab@gmail.com>

Hannah Owens

Andrea Paz <paz.andreita@gmail.com>

Examples

if (FALSE) { # \dontrun{
occs_queryDb(spName = "Bassaricyon alleni", occDb = "gbif", occNum = 10)
} # }