occs_userOccs Loads user provided occurrence records

Load user database with species occurrence records. Returns a list of lists, one per species provided in database in each species list with a set of appropriate fields

Usage

occs_userOccs(txtPath, txtName, txtSep = ",", txtDec = ".", logger = NULL)

Arguments

txtPath

path to database including database name and extension

txtName

name of database without the extension. Database must have at least three columns named 'scientific_name', 'longitude', 'latitude'

txtSep

field separator used in database (as in read.delim)

txtDec

decimal separator used for coordinates in database

logger

Stores all notification messages to be displayed in the Log Window of Wallace GUI. Insert the logger reactive list here for running in shiny, otherwise leave the default NULL

Value

List of lists. One list per species with occurrence records. Each individual species list with appropriate fields for analysis

Details

This function is called by the module occs_queryDb to load a user provided database for species occurrence records, subset to only those records with coordinates, remove records with duplicate coordinates, and select some columns with fields appropriate to studies in biogeography.

Author

Jamie Kass <jamie.m.kass@gmail.com>

Gonzalo E. Pinilla-Buitrago <gepinillab@gmail.com>

Examples

txtPath <- system.file("extdata/Bassaricyon_alleni.csv", package = "wallace")
txtName <- 'Bassaricyon_alleni'
user.occs <- occs_userOccs(txtPath, txtName)
#> <font color="#003300"><b><i>Bassaricyon alleni</i> | </b></font>Data uploaded from <i>'Bassaricyon_alleni'</i>: Duplicated records removed [0]. Remaining records [35].