poccs_removeByID Remove occurrence by ID

This function removes user selected occurrences by ID.

Usage

poccs_removeByID(occs, removeID, logger = NULL, spN = NULL)

Arguments

occs

data frame of cleaned occurrences obtained from component occs: Obtain occurrence data

removeID

the ID of the occurrence to be removed from the occurrences dataframe.

logger

Stores all notification messages to be displayed in the Log Window of Wallace GUI. Insert the logger reactive list here for running in shiny, otherwise leave the default NULL

spN

data frame of cleaned occurrences obtained from component occs: Obtain occurrence data. Used to obtain species name for logger messages.

Value

A new occurence dataframe without the user selected occurrence mantaining all columns from original dataframe for further analyses.

Details

This function is called by the remove occurrences by ID module. It allows for removal of a single occurrence flagged by the user on the map. The function will return a data frame of occurrences with all relevant columns for further analyses and without the occurrence selected by the user.

Author

Jamie Kass <jamie.m.kass@gmail.com>

Gonzalo E. Pinilla-Buitrago <gepinillab@gmail.com>

Examples

occs <- read.csv(system.file("extdata/Bassaricyon_neblina.csv",
                 package = "wallace"))[, 2:3]
occs$occID <- 1:nrow(occs)
out.ID <- poccs_removeByID(occs, 11)
#> Removed occurrence with occID = 11. Updated data has n = 17 records.