Run Wallace Application

This function runs the Wallace application in the user's default web browser.

Usage

run_wallace(launch.browser = TRUE, port = getOption("shiny.port"))

Arguments

launch.browser

Whether or not to launch a new browser window.

port

The port for the shiny server to listen on. Defaults to a random available port.

Note

Please see the official website (https://wallaceecomod.github.io/) for more details. If you have questions about the application, please participate in the Google Group, or email the team directly: wallaceEcoMod@gmail.com.

Author

Jamie Kass <jamie.m.kass@gmail.com>

Gonzalo E. Pinilla-Buitrago <gepinillab@gmail.com>

Examples

if(interactive()) {
run_wallace()
}