Wallace: A modular platform for reproducible ecological modeling

Wallace is a shiny app that guides users through a complete species niche/distributional modeling analysis, from the acquisition of species occurrence and environmental data to visualizing model predictions on an interactive map (rleaflet), thus bundling complex workflows into a single, streamlined GUI interface. New functionality, in the form of modules, can be added to Wallace via contributions from the user community. In addition, executable session code (R Markdown format) can be downloaded to share with others or use as supplementary information for scientific papers and reports. The application is run via the function run_wallace.

Details

Please see the official website (https://wallaceecomod.github.io/) for more details. If you have questions about the application, please participate in the Google Group, or email the team directly: wallaceEcoMod@gmail.com.